Reorganized folders and files. Deleted duplicates and kept most current ones.
