--- app/models/schema.sql
@@ -1,4 +1,0 @@
--- Drops the "ASD_resources" if it exists currently --
-DROP DATABASE IF EXISTS ASD_resources;
--- Creates the "ASD_resources" database --
-CREATE DATABASE ASD_resources;
