--- app/public/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Poisoned Proxies </title>
-    <script src="https://code.jquery.com/jquery.js"></script>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
-        crossorigin="anonymous">
-</head>
-<body>
-
-    <h1>Poisoned Proxies Database Test Page</h1>
-<div class="container">
-
-    <p id="resources-container"></p>
-    
-</div>
-
-<!-- Custom Script -->
-<script src="js/views.js" type="text/javascript"></script>
-   
-
-</body>
-</html>
