+++ models/suggested.js
@@ -2,17 +2,17 @@
 //"DataTypes" defines what type of data
 module.exports = function (sequelize, DataTypes) {
 
-    var ASD_suggested_table = sequelize.define("ASD_suggested_table", {
+    var suggested = sequelize.define("suggested", {
         Name: DataTypes.STRING,
         Location: DataTypes.STRING,
         Phone_Number: DataTypes.STRING,
         Description: DataTypes.TEXT,
         Price: DataTypes.INTEGER,
         Insurance_Accepted: DataTypes.BOOLEAN,
-        Catergory: DataTypes.STRING
+        Category: DataTypes.STRING
     });
 
-    return ASD_suggested_table;
+    return suggested;
 
 };
